requirements.txt
streamlit
boto3
strands-agents
strands-agents-tools
uvx